+++ vinted_api_kit/models/__init__.py
@@ -1,4 +1,0 @@
-from .catalog_item import CatalogItem
-from .detailed_item import DetailedItem
-
-__all__ = ["CatalogItem", "DetailedItem"]